tweetを外に出す
