+++ src/api/app.py
@@ -53,216 +53,20 @@ app.register_blueprint(questions.app, url_prefix="/questions")
 from operations import operations
 app.register_blueprint(operations.app, url_prefix="/operations")
 
+#from informations import informations
+#app.register_blueprint(informations.app, url_prefix="/informations")
+#
+#from responses import responses
+#app.register_blueprint(responses.app, url_prefix="/responses")
+#
+from tweets import tweets
+app.register_blueprint(tweets.app, url_prefix="/tweets")
+
 #@app.route('/')
 #@crossdomain(origin='*')
 #def test():
 #    res = {1: 2}
 #    return jsonify(res)
-
-
-@app.route('/questions', methods=['GET'])
-@crossdomain(origin='*')
-def index_questions():
-    """質問一覧を返却します
-    """
-    code = 200
-    #logging.debug(request.headers)
-    # TODO 公開時にコメントイン
-    #if request.headers['Api-Key'] != API_ACCESS_KEY:
-    #    abort(401)
-
-    questions_dict = {}
-    try:
-        questions = get_questions()
-        logging.debug(questions)
-        questions_dict = ListQuestionMapper({'result': questions}).as_dict()
-    except:
-        pass
-    logging.debug(questions_dict)
-    result = questions_dict['result']
-
-    return jsonify(status_code=code, result=result)
-
-# TODO answerも同時に登録するように修正すること
-@app.route('/questions', methods=['POST'])
-@crossdomain(origin='*')
-def add_question():
-    """リクエストを元に質問を登録します
-    """
-    code = 201
-    tdatetime = dt.now()
-    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
-    req = request.form
-    #logging.debug(req['content'])
-    #logging.debug(req['state'])
-    #logging.debug(req['created_by'])
-    #logging.debug(req['updated_by'])
-    try:
-        question = Question(
-                            id=None,
-                            content=req['content'],
-                            state=req['state'],
-                            created_by=req['created_by'],
-                            updated_by=req['updated_by'],
-                            created_at=tstr,
-                            updated_at=tstr
-        )
-        db_session.add(question)
-        db_session.commit()
-    except:
-        # 登録失敗
-        db_session.rollback()
-    finally:
-        db_session.close()
-
-    return jsonify(status_code=code)
-
-@app.route('/questions/<question_id>', methods=['GET'])
-@crossdomain(origin='*')
-def show_question(question_id):
-    """idを元に質問データを取得します
-    """
-    code = 200
-    question_dict = {}
-
-    try:
-        question = get_question(question_id)
-        question_dict = QuestionMapper(question).as_dict()
-    except:
-        # 取得に失敗
-        pass
-
-    return jsonify(status_code=code, result=question_dict)
-
-def get_question(question_id):
-    """質問idを元にQuestionモデルを返却する
-    """
-    #logging.debug(Question.query.first().id)
-    question = []
-    res = Question.query.filter("id = :question_id").params(question_id=question_id).first()
-    #logging.debug(res)
-    question = Question(id=question_id,
-                        content=res.content,
-                        state=res.state,
-                        created_by=res.created_by,
-                        updated_by=res.updated_by,
-                        created_at=res.created_at,
-                        updated_at=res.updated_at
-    )
-
-    return question
-
-def get_questions():
-    """質問一覧のリストを返却する
-    """
-    questions = []
-    res = Question.query.all()
-    for row in res:
-        # get_question(row["id"])でmodelをセットしていく？
-        questions.append(row)
-
-    return questions
-
-@app.route('/questions/<question_id>', methods=['PUT'])
-@crossdomain(origin='*')
-def edit_question(question_id):
-    """質問を更新します
-    """
-    code = 201
-    tdatetime = dt.now()
-    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
-    req = request.form
-    try:
-        row = db_session.query(Question).get(question_id)
-        row.content = req['content']
-        row.state = req['state']
-        row.created_by = req['created_by']
-        row.updated_by = req['updated_by']
-        row.updated_at = tstr
-        db_session.flush()
-        #db_session.commit()
-    except:
-        pass
-    finally:
-        pass
-
-    # app.logger.debug(request.form)
-    return jsonify(status_code=code)
-
-@app.route('/questions/<question_id>', methods=['DELETE'])
-@crossdomain(origin='*')
-def delete_question(question_id):
-    """質問を削除します
-    """
-    code = 204
-    try:
-        row = Question.query.get(question_id)
-        db_session.delete(row)
-        db_session.flush()
-    except:
-        pass
-    finally:
-        pass
-
-    #logging.debug(row.created_by)
-
-    return jsonify(status_code=code)
-
-
-# /answers
-@app.route('/answers/<question_id>', methods=['POST'])
-@crossdomain(origin='*')
-def add_answer(question_id):
-    code = 201
-    tdatetime = dt.now()
-    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
-    g.cursor.execute('insert into answers (question_id, content, state, created_by, updated_by, created_at, updated_at) values (%s, %s, %s, %s, %s, %s, %s)',
-            ([question_id, request.form['content'], request.form['state'], request.form['created_by'], request.form['updated_by'], tstr, tstr]))
-    g.connection.commit()
-
-    return jsonify(status_code=code)
-
-@app.route('/questions/<question_id>/answers', methods=['GET'])
-@crossdomain(origin='*')
-def index_answers(question_id):
-    code = 200
-    g.cursor.execute('select id, question_id, content, state, created_by, updated_by, created_at, updated_at from answers where question_id = %s',
-            question_id)
-    answers = [dict(id=row[0], question_id=row[1], content=row[2], state=row[3], created_by=row[4], updated_by=row[5], created_at=row[6], updated_at=row[7]) for row in g.cursor.fetchall()]
-
-    return jsonify(status_code=code, result=answers)
-
-@app.route('/answers/<answer_id>', methods=['GET'])
-@crossdomain(origin='*')
-def show_answer(answer_id):
-    code = 200
-    g.cursor.execute('select id, question_id, content, state, created_by, updated_by, created_at, updated_at from answers where id = %s',
-            [answer_id])
-    question = [dict(id=row[0], question_id=row[1], content=row[2], state=row[3], created_by=row[4], updated_by=row[5], created_at=row[6], updated_at=row[7]) for row in g.cursor.fetchall()]
-
-    return jsonify(status_code=code, result=question)
-
-@app.route('/answers/<answer_id>', methods=['PUT'])
-@crossdomain(origin='*')
-def edit_answer(answer_id):
-    code = 201
-    tdatetime = dt.now()
-    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
-    g.cursor.execute('update answers set content = %s, state = %s, updated_by = %s, updated_at = %s where id = %s',
-            ([request.form['content'], request.form['state'], request.form['updated_by'], tstr, answer_id]))
-    g.connection.commit()
-
-    return jsonify(status_code=code)
-
-@app.route('/answers/<answer_id>', methods=['DELETE'])
-@crossdomain(origin='*')
-def delete_answer(answer_id):
-    code = 204
-    g.cursor.execute('delete from answers where id = %s',
-    [answer_id])
-    g.connection.commit()
-
-    return jsonify(status_code=code)
 
 @app.route('/informations', methods=['POST'])
 @crossdomain(origin='*')
@@ -392,37 +196,6 @@ def delete_information(information_id):
 
     return jsonify(status_code=code)
 
-@app.route('/tweets', methods=['GET'])
-@crossdomain(origin='*')
-def index_tweets():
-    """ツイート一覧を返します
-    """
-    code = 200
-    tweets = []
-
-    try:
-        tweets = get_tweets()
-        tweets_dict = ListTweetMapper({'result': tweets}).as_dict()
-    except:
-        pass
-
-    result = tweets_dict['result']
-
-    return jsonify(status_code=code, result=result)
-
-# TODO pagingを実装する
-def get_tweets():
-    """dbからツイートを全件取得します
-    """
-    tweets = []
-    res = Tweet.query.all()
-    for row in res:
-        tweets.append(row)
-
-    return tweets
-
-# TODO tweetを追加するapiの実装
-# TODO filterの実装
 
 @app.route('/responses', methods=['POST'])
 @crossdomain(origin='*')
